RenderClient can now provide the example image for the TTStyleDelegate protocol.
Just drag-and-drop your image from the Finder into the image well on the left-side of the RenderClient. Then click the "Submit Configuration" button to upload the image and request a new rasterization.

I also updated the README to describe how to use the client with respect to the TTStyleDelegate protocol.